feat(agenthub): add shared schemas and utils
