+++ back/agenthub/auth/schemas.py
@@ -1,24 +1,5 @@
-# ============================================
-# back/agenthub/auth/schemas.py - SIMPLIFICADO
-# ============================================
+"""Legacy import file for authentication schemas."""
 
-from pydantic import BaseModel, EmailStr
+from agenthub.schemas import UserCreate, UserOut, SignInInput, TokenResponse
 
-class UserCreate(BaseModel):
-    email: EmailStr
-    password: str
-
-class UserOut(BaseModel):
-    id: int
-    email: EmailStr
-    is_active: bool
-    
-    # Configuración para Pydantic V2
-    model_config = {"from_attributes": True}
-
-class SignInInput(BaseModel):
-    email: EmailStr
-    password: str
-
-class TokenResponse(BaseModel):
-    token: str
+__all__ = ["UserCreate", "UserOut", "SignInInput", "TokenResponse"]
