+++ back/agenthub/schemas.py
@@ -1,0 +1,54 @@
+"""Common Pydantic models used across AgentHub."""
+
+from pydantic import BaseModel, EmailStr
+
+
+class UserCreate(BaseModel):
+    """Payload for creating a new user."""
+
+    email: EmailStr
+    password: str
+
+
+class UserOut(BaseModel):
+    """Representation of a user returned by the API."""
+
+    id: int
+    email: EmailStr
+    is_active: bool
+
+    # Pydantic v2 configuration
+    model_config = {"from_attributes": True}
+
+
+class SignInInput(BaseModel):
+    """Credentials for user login."""
+
+    email: EmailStr
+    password: str
+
+
+class TokenResponse(BaseModel):
+    """Response returned after successful authentication."""
+
+    access_token: str
+    token_type: str = "bearer"
+
+
+class OAuthUser(BaseModel):
+    """Basic data returned by external OAuth providers."""
+
+    id: str
+    email: str
+    name: str
+    avatar_url: str | None = None
+    provider: str
+
+
+__all__ = [
+    "UserCreate",
+    "UserOut",
+    "SignInInput",
+    "TokenResponse",
+    "OAuthUser",
+]
