+++ back/agenthub/utils.py
@@ -1,19 +1,32 @@
-# agenthub/logging_config.py
+"""Utility helpers used across AgentHub."""
+
+from __future__ import annotations
+
 import logging
 import logging.handlers
+import os
 from pathlib import Path
 
 from agenthub.config import config
 
+# ---------------------------------------------------------------------------
+# Common constants
+# ---------------------------------------------------------------------------
 
-def setup_logging():
-    """Configura el sistema de logging"""
+SECRET_KEY = os.getenv("AGENTHUB_SECRET_KEY", config.get("secret_key", "dev-secret-key"))
+ALGORITHM = "HS256"
+ACCESS_TOKEN_EXPIRE_MINUTES = 30
 
-    # Crear directorio de logs
+# ---------------------------------------------------------------------------
+# Logging utilities
+# ---------------------------------------------------------------------------
+
+def setup_logging() -> None:
+    """Configure application-wide logging."""
+
     log_dir = Path("logs")
     log_dir.mkdir(exist_ok=True)
 
-    # Configuración de formato
     formatter = logging.Formatter(
         fmt=config.get("logging", {}).get(
             "format", "%(asctime)s - %(name)s - %(levelname)s - %(message)s"
@@ -21,23 +34,19 @@ def setup_logging():
         datefmt="%Y-%m-%d %H:%M:%S",
     )
 
-    # Logger raíz
     root_logger = logging.getLogger()
     root_logger.setLevel(getattr(logging, config.get("log_level", "INFO")))
 
-    # Handler para consola
     console_handler = logging.StreamHandler()
     console_handler.setFormatter(formatter)
     root_logger.addHandler(console_handler)
 
-    # Handler para archivo
     file_handler = logging.handlers.RotatingFileHandler(
-        log_dir / "agenthub.log", maxBytes=10 * 1024 * 1024, backupCount=5  # 10MB
+        log_dir / "agenthub.log", maxBytes=10 * 1024 * 1024, backupCount=5
     )
     file_handler.setFormatter(formatter)
     root_logger.addHandler(file_handler)
 
-    # Handler para errores
     error_handler = logging.handlers.RotatingFileHandler(
         log_dir / "errors.log", maxBytes=10 * 1024 * 1024, backupCount=3
     )
@@ -45,6 +54,13 @@ def setup_logging():
     error_handler.setFormatter(formatter)
     root_logger.addHandler(error_handler)
 
-    # Configurar loggers específicos
     logging.getLogger("uvicorn").setLevel(logging.WARNING)
     logging.getLogger("fastapi").setLevel(logging.INFO)
+
+
+__all__ = [
+    "setup_logging",
+    "SECRET_KEY",
+    "ALGORITHM",
+    "ACCESS_TOKEN_EXPIRE_MINUTES",
+]
